move block builder to preflight crate
