+++ bin/zeth/Cargo.toml
@@ -19,7 +19,6 @@ workspace = true
 [dependencies]
 alloy.workspace = true
 anyhow.workspace = true
-async-trait.workspace = true
 bincode.workspace = true
 bytemuck.workspace = true
 clap.workspace = true
@@ -28,8 +27,6 @@ log.workspace = true
 pot.workspace = true
 
 reth-chainspec.workspace = true
-reth-primitives.workspace = true
-reth-revm.workspace = true
 
 serde.workspace = true
 tokio.workspace = true
