+++ crates/preflight/Cargo.toml
@@ -9,6 +9,7 @@ workspace = true
 [dependencies]
 alloy.workspace = true
 anyhow.workspace = true
+async-trait.workspace = true
 env_logger.workspace = true
 hashbrown.workspace = true
 log.workspace = true
