+++ crates/preflight/src/driver.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use alloy::primitives::U256;
+use alloy::primitives::{B256, U256};
 use alloy::rpc::types::{Block, Header};
 use zeth_core::stateless::driver::SCEDriver;
 
@@ -20,6 +20,14 @@ use zeth_core::stateless::driver::SCEDriver;
 pub struct AlloyDriver;
 
 impl SCEDriver<Block, Header> for AlloyDriver {
+    fn header_hash(header: &Header) -> B256 {
+        header.hash
+    }
+
+    fn block_header(block: &Block) -> &Header {
+        &block.header
+    }
+
     fn block_to_header(block: Block) -> Header {
         block.header
     }
